backend/requirements.txt
connexion[flask]==3.0.6
PyYAML==6.0.1
SQLAlchemy==2.0.30
PyJWT==2.8.0
bcrypt==4.1.2
Pillow==10.3.0
